一些bug修复
ClientCore中有个关键的bug修复：catch (Exception ex) when (activity is not null)
增加 ReasoningContent

+++ dotnet/src/Plugins/Plugins.Web/IRelevantContent.cs
@@ -4,5 +4,5 @@ namespace Microsoft.SemanticKernel.Plugins.Web;
 public interface IRelevantContent
 {
     public string Text { get; }
-    public double Relevance { get; set; }
+    public float Relevance { get; set; }
 }
